pyproject.toml
[project]
name = "cybersec-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.1",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
]
